requirements.txt
streamlit
pillow
gspread
oauth2client
black